- fix Windows builds, cast int to size_t.
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@15967 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ SimulationRuntime/c/simulation/simulation_runtime.cpp
@@ -123,7 +123,7 @@ static void setTermMsg(const char *msg, va_list ap)
   size_t i;
   static size_t termMsgSize = 0;
   if (TermMsg==NULL) {
-    termMsgSize = max(strlen(msg)*2+1,2048);
+    termMsgSize = max(strlen(msg)*2+1,(size_t)2048);
     TermMsg = (char*) malloc(termMsgSize);
   }
   i = vsnprintf(TermMsg,termMsgSize,msg,ap);
